Playground styling updates (#2259)
