+++ packages/client/package.json
@@ -42,6 +42,7 @@
         "react-hook-form": "^7.34.2",
         "react-router-dom": "^6.3.0",
         "react-color": "^2.19.3",
+        "flexlayout-react": "^0.8.1",
         "prettier": "^2.4.1",
         "tailwindcss": "^4.1.14"
     },
