+++ packages/client/src/components/blocks-workspace/panels/LayersPanel.tsx
@@ -33,7 +33,6 @@ import {
 	INPUT_BLOCK_TYPES,
 	useBlocks,
 } from "@semoss/renderer";
-import { FlexLayout } from "@semoss/shared";
 import {
 	Divider,
 	Grid,
@@ -48,6 +47,7 @@ import {
 	Typography,
 	useNotification,
 } from "@semoss/ui";
+import { FlexLayout } from "@/components/flex-layout";
 import { AddVariableModal } from "@/components/notebook";
 import { Panel } from "@/components/workspace";
 import { useDesigner, useWorkspace } from "@/hooks";
