+++ packages/client/src/components/flex-layout/flexlayout.css
@@ -1,0 +1,158 @@
+/*
+  Layout
+*/
+.flexlayout__layout {
+    --color-background: transparent;
+    --font-family: 'Inter', sans-serif;
+    --font-size: 0.875rem;
+    --smss-tab-border-radius: 4px;
+}
+.flexlayout__border_left {
+    width: 46px;
+    margin-right: 10px;
+    border-radius: 12px;
+    background: var(--Background-Paper1, #fff);
+    box-shadow: 0px 1px 2px rgba(0, 0, 0, 0.08);
+    border-right: none;
+}
+.flexlayout__border_button--unselected {
+    background-color: #fff;
+}
+/*Should take another look into this*/
+.flexlayout__border_inner_tab_container.flexlayout__border_inner_tab_container_right{
+    top: 6px !important;
+    left: 84% !important;
+}
+.flexlayout__border.flexlayout__border_right{
+    width: 40px;
+    border-radius: var(--Shape-borderRadiusLg, 12px);
+    background: var(--Background-Paper1, #FFF);
+    margin-left: 8px;
+    border-left: none;
+}
+
+.flexlayout__border_inner_tab_container {
+    padding-left: 0px;
+    padding-right: 0px;
+}
+
+.flexlayout__border_left .flexlayout__border_button {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    align-self: stretch;
+    transform: rotate(90deg);
+    padding: 0;
+}
+
+.flexlayout__border_button_right {
+    padding: 3px 0.5em;
+    border-radius: 16px;
+    background: var(--Primary-Selected, #ebf4fe);
+    color: var(--Primary-Dark, #1260dd);
+    font-size: 13px;
+    line-height: 18px; /* 138.462% */
+}
+
+.flexlayout__border_left .flexlayout__border_button--selected {
+    background-color: rgb(255, 255, 255);
+    color: rgb(4, 113, 240);
+}
+
+.flexlayout__border_bottom {
+    border-top: none;
+    padding-left: 34px;
+}
+
+.workspace_layout .flexlayout__border_bottom {
+    padding-left: 0;
+    position: absolute;
+    bottom: 0;
+    min-width: 90px;
+}
+
+.workspace_layout .flexlayout__border_bottom .flexlayout__border_button_bottom {
+   background: transparent;
+}
+
+/* 
+Main Content tabs
+*/
+
+.flexlayout__tabset_tabbar_inner_tab_container_top {
+    border: none;
+}
+
+.flexlayout__tabset_tab_divider {
+    width: 0px;
+}
+.flexlayout__tab_moveable {
+    margin-left: 5px;
+}
+.flexlayout__tab.flexlayout__tab_border.flexlayout__tab_border_right {
+    border-radius: var(--Shape-borderRadiusLg, 12px);
+    background: var(--Background-Paper1, #FFF);
+}
+.flexlayout__tab {
+    background: transparent;
+    border-radius: var(--smss-tab-border-radius);
+    border-top-left-radius: 0;
+}
+.selected_block.flexlayout__border_bottom {
+    border-top: 1px solid rgba(0, 0, 0, 0.12);
+    background: var(--Background-Paper1, #fff);
+    box-shadow: 0px 1px 2px rgba(212, 12, 12, 0.76);
+}
+
+.flexlayout__tabset_tabbar_outer_top {
+    border: none;
+}
+
+.flexlayout__tabset_tabbar_inner_tab_container {
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    flex: 1 0 0;
+}
+.flexlayout__tabset_container{
+    border-radius: var(--Shape-borderRadiusLg, 12px);
+    background: var(--Background-Paper1, #fff);
+}
+.flexlayout__tabset-selected {
+    background-color: transparent;
+}
+.flexlayout__tabset {
+    margin-top: 8px;
+    margin-left: 4px;
+}
+.flexlayout__tab.flexlayout__tab_border.flexlayout__tab_border_left {
+    margin-left: 0px;
+    border-radius: var(--Shape-borderRadiusLg, 12px);
+    background: var(--Background-Paper1, #FFF);
+    /* Elevation - SEMOSS/1 */
+    box-shadow: 0px 1px 2px 0px rgba(0, 0, 0, 0.08);
+}
+.flexlayout__tab_button {
+    display: flex;
+    align-items: center;
+    border-radius: 16px;
+    background: var(--Secondary-Selected, #ebebeb);
+    font-size: 13px;
+    line-height: 18px; /* 138.462% */
+}
+
+.flexlayout__tab_button:hover {
+    background-color: rgba(0, 0, 0, 0.04);
+    background-color: var(--Secondary-Selected, #ebf4fe);
+}
+
+.flexlayout__tab_button_trailing svg {
+    width: 1rem;
+    height: 1rem;
+    fill: var(--color-icon, #000);
+}
+
+.flexlayout__tab_button--selected {
+    color: var(--Primary-Dark, #1260dd);
+    background-color: #ebf4fe;
+}
