+++ packages/client/src/components/flex-layout/index.ts
@@ -1,0 +1,6 @@
+// import custom styling
+import "flexlayout-react/style/light.css";
+import "./flexlayout.css";
+
+// export from lib
+export * as FlexLayout from "flexlayout-react";
