+++ packages/client/src/components/workspace/Workspace.tsx
@@ -3,7 +3,6 @@ import { observer } from "mobx-react-lite";
 import type React from "react";
 import { useEffect, useRef } from "react";
 import { Link } from "react-router-dom";
-import { FlexLayout } from "@semoss/shared";
 import {
 	Breadcrumbs,
 	IconButton,
@@ -14,6 +13,7 @@ import {
 } from "@semoss/ui";
 import { ClosePage } from "@/assets/img/ClosePage";
 import SEMOSS_BLACK_LOGO from "@/assets/img/SEMOSS_BLACK_LOGO.png";
+import { FlexLayout } from "@/components/flex-layout";
 import { WorkspaceContext } from "@/contexts";
 import { SIDEBAR_MENU } from "@/pages/import/import.constants";
 import type { WorkspaceOptions, WorkspaceStore } from "@/stores";
