+++ packages/client/src/stores/workspace/workspace.types.ts
@@ -1,4 +1,4 @@
-import type { FlexLayout } from "@semoss/shared";
+import type { FlexLayout } from "@/components/flex-layout";
 
 export interface WorkspaceOptions {
 	version: string;
