+++ packages/playground/src/components/room/room-tool.tsx
@@ -126,7 +126,6 @@ export const RoomTool: React.FC<RoomToolProps> = observer(({ node, room }) => {
 
 	useEffect(() => {
 		const checkPortal = async () => {
-			console.log({ config });
 			// Finish loading
 			if (
 				getAppInfo.status === "INITIAL" ||
@@ -180,8 +179,6 @@ export const RoomTool: React.FC<RoomToolProps> = observer(({ node, room }) => {
 	if (!config) {
 		return <div>No Tool</div>;
 	}
-
-	console.log({ isLoading, url, selectedTool });
 
 	return (
 		<div className="relative flex h-full w-full flex-col items-center justify-center overflow-hidden">
